Add log session identifier to output messages
Summary:
Add log session identifier to output messages.  This is a
unique identifier based on time in microseconds that applies
to a sshd process and its children.  It is intended to make
sessionization of ssh logs simpler by having a single key
that works on all log lines.  It works by setting an
environment variable which children can check.  A PAM
environment variable is also set, which PAM processes
need to check and add to their own syslog messages.

Differential Revision: D18308744

fbshipit-source-id: c5a7431cfac9101a0c2276461b653dcfbc62515c