Pass the full principal list of an ssh client certificate to child process
Summary:
This patch sets an environment variable SSH_CERT_PRINCIPALS in the child
process to be the full principal list of a user's SSH certificate when forced
command is present and the user is authenticated by the certificate. When
there is more than one certificate in a certificate file, SSH_CERT_PRINCIPALS
is set to the list of principals in the first cert that matches a principal in
the authorized_principals file.

The principals are separated by commas. Comma is used by ssh-keygen when
creating an SSH certificate with a list of principals.

Every line in an authorized_keys file is limited to be SSH_MAX_PUBKEY_BYTES
bytes. The certiciate file is also a type of authorized_keys file. Each
certificate is a single line in a certificate file and is limited to be
SSH_MAX_PUBKEY_BYTES bytes.

Differential Revision: D19134213

fbshipit-source-id: 9933ece6