Remove HPN patch
Summary:
In the release from OpenSSH 7.5->7.6 the HPN patch suffered a big
regression in throughput. Testing shows that without HPN OpenSSH now
actually has higher throughput on high latency connections (150ms+).

There is therefore no more use for HPN, which in fact is now having a
negative impact on throughput, so remove it.

Reviewed By: RichardScothern

Differential Revision: D25120306

fbshipit-source-id: 802d6c6c