fix build script and add hpn patches back temporarily.
Summary: This change fixes the build-srpm.sh script for the 8.0 branch, updates the openssh.spec file, and adds back the HPN patch so that we don't have to make a lot of changes to how ssh is configured everywhere for the LOG_SESSION_ID change push. There seems to have been a build created for openssh 8.0 without the HPN patch, but it doesn't seem to have ever been deployed (because the latest rpm version that is out there doesn't seem to be configured to be installed anywhere.). On the part of fixing the build-srpm.sh script, I updated the pax pattern so that it would prepend the needed folder name (openssh-$ver-$rel) and a slash to all file names passed to pax instead of  the old pattern which was replacing the first . in a filename with "openssh-$ver-$rel".

Reviewed By: xuhang57

Differential Revision: D31057335

fbshipit-source-id: a3ece03b